backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
cryptography==45.0.6
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.11.0
django-cors-headers==4.7.0
django-email-users==2.1.0
django-filter==25.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
idna==3.10
iniconfig==2.1.0
numpy==1.24.3
packaging==25.0
pandas==1.5.3
phonenumberslite==9.0.12
Pillow==11.3.0
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.4
scipy==1.10.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0